Refactor: move `create_encrypted_to_device_request` in Device
